+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/magic/TriggerType.java
@@ -1,0 +1,15 @@
+package com.elmakers.mine.bukkit.api.magic;
+
+public enum TriggerType {
+    // MOB-ONLY
+    SPAWN,
+
+    // PLAYER-ONLY
+    JOIN,
+
+    // COMMON
+    INTERVAL, DEATH, DAMAGE, LAUNCH,
+
+    // TODO, not working yet
+    DEAL_DAMAGE
+}