+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/spell/Spell.java
@@ -108,8 +108,6 @@ public interface Spell extends SpellTemplate {
      * @return true if the spell was in the middle of selection and was cancelled.
      */
     boolean cancelSelection();
-    boolean isPassive();
     boolean isEnabled();
     void setEnabled(boolean enabled);
-    boolean isToggleable();
 }