+++ Magic/CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG
 
 # 7.9.2
-
- - Add "passive" flag for spells that can not be directly cast but still show up in the spell inventory
+ 
+ - Passive spell system!
+   - Add "triggers" option to a spell, it can be auto-cast under various circumstances
+   - Add "passive" flag for spells that can not be directly cast but still show up in the spell inventory
+   - See the "savior" spell in the survival configs for an example of use
 
 # 7.9.1